pyproject.toml
[project]
name = "resqdb"
version = "0.1.0"
description = "Package to upload ResQEnergy data to a database"
requires-python = ">=3.13"
dependencies = [
    "faker>=38.0.0",
    "geoalchemy2>=0.18.0",
    "geopandas>=1.1.1",
    "loguru>=0.7.3",
    "oemof-tabular",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]

[tool.uv.sources]
oemof-tabular = { git = "https://github.com/oemof/oemof-tabular" }

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
]

# ==== ruff ====
[tool.ruff]
output-format = "concise"  # This enables file-links to errors in git console
line-length = 120
lint.select = ["ALL"]
exclude = [
  "manage.py",
  "slapp/utils/ogr_layer_mapping.py",
  "config/wsgi.py",
  "slapp/contrib/*",
  "slapp/users/*",
  "merge_local_dotenvs_in_dotenv.py",
  "slapp/utils/context_processors.py"
]
lint.ignore = [
  "I001",    # Import block is un-sorted or un-formatted (done by isort)
  "D203",    # 1 blank line required before class docstring
  "D212",    # Multi-line docstring summary should start at the first line pydocstyle
  "ANN003",  # Missing type annotation for `**kwargs`
  "EM102",   # Exception must not use an f-string literal, assign to variable first
  "TRY003",  # Avoid specifying long messages outside the exception class
  "S101",    # Use of `assert` detected
  "UP007",   # Use `X | Y` for type annotations
  "B905",    # `zip()` without an explicit `strict=` parameter
  "FIX001",  # Line contains FIXME
  "FIX002",  # Line contains TODO
  "RET504",  # Unnecessary variable assignment before `return` statement
  "G004",    # Logging statement uses f-string
  "PD011",   # Use `.to_numpy()` instead of `.values`  (does not work out of the box)
  "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
]
fix = true
show-fixes = true
lint.unfixable = ["UP007", "I001"]
